chatbot/requirements.txt
# AI Agent Dependencies
# Phase III Step 4: OpenAI Agents SDK Integration

# Core agent framework
openai-agents>=0.7.0

# Note: ChatKit backend is implemented using custom classes (chatbot/chatkit/)
# that bridge to the OpenAI Agents SDK. No external chatkit package needed.

# Schema validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# API framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP client for MCP communication
httpx>=0.26.0

# Async support
anyio>=4.0.0

# Date/time handling
python-dateutil>=2.8.0

# JWT authentication (shared with backend)
pyjwt>=2.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0  # for TestClient

# Database (PostgreSQL/Neon)
sqlmodel>=0.0.16
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Development
python-dotenv>=1.0.0
